requirements.txt
Flask==2.3.2
python-docx==0.8.11
openpyxl==3.1.2
PyPDF2==3.0.1
docx2pdf==0.1.8
customtkinter==5.2.0